+++ onmt/dynamic/iterator.py
@@ -7,7 +7,7 @@ from onmt.inputters import str2sortkey
 from onmt.inputters.inputter import max_tok_len, OrderedIterator
 
 # from onmt.dynamic.shard import get_corpora_shards, build_sharded_corpora_iters
-from onmt.dynamic.shard import get_corpora, build_corpora_iters
+from onmt.dynamic.corpus import get_corpora, build_corpora_iters
 from onmt.dynamic.transform import load_transforms
