+++ onmt/dynamic/preprocess.py
@@ -3,7 +3,7 @@ from onmt.utils.logging import init_logger
 
 from onmt.dynamic.parse import DynamicArgumentParser
 from onmt.dynamic.opts import dynamic_preprocess_opts
-from onmt.dynamic.shard import save_dataset, save_transformed_sample
+from onmt.dynamic.corpus import save_dataset, save_transformed_sample
 from onmt.dynamic.vocab import build_dynamic_fields, save_fields
 from onmt.dynamic.transform import make_transforms, save_transforms, \
     get_specials, get_transforms_cls
